almost done with first-pass arduino intergration (not tested)
