+++ devices/arduino/README.md
@@ -1,3 +1,5 @@
 ##Arduino Sketches
-Each subdirectory contains arduino .ino (and possibly .cpp/.c) files that, when
-compiled and loaded onto a an arduino, will interface with python RobotDevices.
+Each arduino device (or class of devices) has its own .py file and directory 
+containing the arduino sketch and other files that, when complied and loaded
+onto an arduino, allow it to interface with the associated python 
+ArduinoRobotDevice instance.
